Move DockingResultCols enum from docking to data package
`DockingResultCols` was defined in `asapdiscovery.docking.docking_data_validation`
but imported at runtime by `asapdiscovery.data.services.postera.manifold_artifacts`,
creating a `data` -> `docking` runtime dependency that undermines `data`'s role
as the dependency-free foundation package.

Since it is just a string enum of column names with no docking logic, it now
lives in `asapdiscovery.data.schema.docking_data_validation`.

- Add `DockingResultCols` to the data package
- Replace the docking module with a backwards-compatible re-export so existing
  `asapdiscovery.docking.docking_data_validation` imports keep working
- Repoint all internal imports (data, dataviz, docking, workflows) to the new
  data-package location, breaking the data -> docking edge

Closes #1292
