+++ asapdiscovery-docking/asapdiscovery/docking/tests/test_docking.py
@@ -3,7 +3,7 @@ from pathlib import Path
 
 import pytest
 
-from asapdiscovery.docking.docking_data_validation import DockingResultCols
+from asapdiscovery.data.schema.docking_data_validation import DockingResultCols
 from asapdiscovery.docking.openeye import POSIT_METHOD, POSITDocker
