+++ asapdiscovery-workflows/asapdiscovery/workflows/docking_workflows/cross_docking.py
@@ -20,7 +20,7 @@ from asapdiscovery.docking.docking import (
     DockingInputMultiStructure,
     write_results_to_multi_sdf,
 )
-from asapdiscovery.docking.docking_data_validation import DockingResultCols
+from asapdiscovery.data.schema.docking_data_validation import DockingResultCols
 from asapdiscovery.docking.openeye import POSIT_METHOD, POSIT_RELAX_MODE, POSITDocker
 from asapdiscovery.docking.scorer import ChemGauss4Scorer, MetaScorer
 from asapdiscovery.modeling.protein_prep import ProteinPrepper
